ref_name: move RemoteRefSymbol types from refs module
