+++ cli/src/git_util.rs
@@ -38,7 +38,7 @@ use jj_lib::git::GitImportStats;
 use jj_lib::git::GitRefKind;
 use jj_lib::op_store::RefTarget;
 use jj_lib::op_store::RemoteRef;
-use jj_lib::refs::RemoteRefSymbol;
+use jj_lib::ref_name::RemoteRefSymbol;
 use jj_lib::repo::ReadonlyRepo;
 use jj_lib::repo::Repo;
 use jj_lib::workspace::Workspace;
