+++ lib/src/ref_name.rs
@@ -286,3 +286,90 @@ impl Display for RefSymbol {
         f.pad(&revset::format_symbol(&self.0))
     }
 }
+
+/// Owned remote bookmark or tag name.
+///
+/// This type can be displayed in `{name}@{remote}` form, with quoting and
+/// escaping if necessary.
+#[derive(Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
+pub struct RemoteRefSymbolBuf {
+    /// Local name.
+    pub name: String,
+    /// Remote name.
+    pub remote: String,
+}
+
+impl RemoteRefSymbolBuf {
+    /// Converts to reference type.
+    pub fn as_ref(&self) -> RemoteRefSymbol<'_> {
+        RemoteRefSymbol {
+            name: &self.name,
+            remote: &self.remote,
+        }
+    }
+}
+
+/// Borrowed remote bookmark or tag name.
+///
+/// This type can be displayed in `{name}@{remote}` form, with quoting and
+/// escaping if necessary.
+#[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
+pub struct RemoteRefSymbol<'a> {
+    /// Local name.
+    pub name: &'a str,
+    /// Remote name.
+    pub remote: &'a str,
+}
+
+impl RemoteRefSymbol<'_> {
+    /// Converts to owned type.
+    pub fn to_owned(self) -> RemoteRefSymbolBuf {
+        RemoteRefSymbolBuf {
+            name: self.name.to_owned(),
+            remote: self.remote.to_owned(),
+        }
+    }
+}
+
+impl From<RemoteRefSymbol<'_>> for RemoteRefSymbolBuf {
+    fn from(value: RemoteRefSymbol<'_>) -> Self {
+        value.to_owned()
+    }
+}
+
+impl PartialEq<RemoteRefSymbol<'_>> for RemoteRefSymbolBuf {
+    fn eq(&self, other: &RemoteRefSymbol) -> bool {
+        self.as_ref() == *other
+    }
+}
+
+impl PartialEq<RemoteRefSymbol<'_>> for &RemoteRefSymbolBuf {
+    fn eq(&self, other: &RemoteRefSymbol) -> bool {
+        self.as_ref() == *other
+    }
+}
+
+impl PartialEq<RemoteRefSymbolBuf> for RemoteRefSymbol<'_> {
+    fn eq(&self, other: &RemoteRefSymbolBuf) -> bool {
+        *self == other.as_ref()
+    }
+}
+
+impl PartialEq<&RemoteRefSymbolBuf> for RemoteRefSymbol<'_> {
+    fn eq(&self, other: &&RemoteRefSymbolBuf) -> bool {
+        *self == other.as_ref()
+    }
+}
+
+impl Display for RemoteRefSymbolBuf {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        Display::fmt(&self.as_ref(), f)
+    }
+}
+
+impl Display for RemoteRefSymbol<'_> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let RemoteRefSymbol { name, remote } = self;
+        f.pad(&revset::format_remote_symbol(name, remote))
+    }
+}
