+++ lib/src/repo.rs
@@ -73,11 +73,11 @@ use crate::op_store::RemoteRefState;
 use crate::op_store::RootOperationData;
 use crate::op_store::WorkspaceId;
 use crate::operation::Operation;
+use crate::ref_name::RemoteRefSymbol;
 use crate::refs::diff_named_ref_targets;
 use crate::refs::diff_named_remote_refs;
 use crate::refs::merge_ref_targets;
 use crate::refs::merge_remote_refs;
-use crate::refs::RemoteRefSymbol;
 use crate::revset;
 use crate::revset::RevsetExpression;
 use crate::revset::RevsetIteratorExt as _;
