+++ lib/src/revset.rs
@@ -46,8 +46,8 @@ use crate::object_id::PrefixResolution;
 use crate::op_store::RemoteRefState;
 use crate::op_store::WorkspaceId;
 use crate::op_walk;
-use crate::refs::RemoteRefSymbol;
-use crate::refs::RemoteRefSymbolBuf;
+use crate::ref_name::RemoteRefSymbol;
+use crate::ref_name::RemoteRefSymbolBuf;
 use crate::repo::ReadonlyRepo;
 use crate::repo::Repo;
 use crate::repo::RepoLoaderError;
